merge

+++ include/myst/bits.h
@@ -12,7 +12,7 @@ MYST_INLINE bool myst_test_bit(const uint8_t* data, size_t index)
 {
     const size_t byte = index / 8;
     const size_t bit = index % 8;
-    return ((size_t)(data[byte]) & (size_t)(1 << bit)) ? 1 : 0;
+    return (data[byte] & (uint8_t)(1 << bit)) ? 1 : 0;
 }
 
 MYST_INLINE void myst_set_bit(uint8_t* data, size_t index)
